Add runtime migrations logic
