--- docker-migrations/migrate.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-cd ./docker-migrations
-echo "*** Applying Migrations ***"
-pnpm tsx migrate.ts
-
-exec "$@"
-
